[Numerics] Add the possibility to choose the NSN formulation inside the NSGS solver

+++ Numerics/src/NonSmoothSolvers/FrictionContact/GlobalFrictionContact3D_AlartCurnier.c
@@ -483,9 +483,9 @@ void globalFrictionContact3D_AlartCurnier(
 
   unsigned int localProblemSize = problem->H->size1;
 
-  assert(localProblemSize == problem->numberOfContacts * problem->dimension);
+  assert((int)localProblemSize == problem->numberOfContacts * problem->dimension);
 
-  assert(globalProblemSize == problem->H->size0); /* size(velocity) ==
+  assert((int)globalProblemSize == problem->H->size0); /* size(velocity) ==
                                                    * Htrans*globalVelocity */
